backend/requirements.txt
# Backend Service Requirements
# Install with: pip install -r backend_requirements.txt

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data processing
numpy==1.26.3
pandas==2.2.0

# HTTP clients
httpx==0.26.0
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging and monitoring
python-json-logger==2.0.7

# Async support
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
